Create ComputerBackup
This is a PowerShell script to backup directories on my computer. The paths are stored in a text file. The program is supposed to copy over files that do not already exist and compare MD5 HASH keys on existing files. If the file is modified, the source overwrites the destination.